+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added reader for LCModel RAW format data.
 * Added read_mri_dyn_dir function for reading dynamic MRS exported from Siemens
 scanners.
+* Bug fixes for 2D MRSI voxel segmentation calculation.
 
 # spant 0.19.0
 * Added Asc, BHB, Cho, PEth and Ser simulation parameters.