+++ NEWS.md
@@ -16,7 +16,7 @@ scanners.
 * Added a function to grid shift 2D MRSI data in the x/y direction.
 * Better plotting/fitting support for masking data by setting data points to NA.
 * Bug fix for interactive voxel selection position indicator.
-* Added mask_xy to mask voxels in a centered rectangular region.
+* Added mask_xy to mask voxels in a centred rectangular region.
 * Minor changes to improve parallel processing support.
 
 # spant 0.16.0